+++ tests/service/views/test_version_views.py
@@ -26,9 +26,13 @@ def test_version(svc_client):
     assert "result" in response.json
     data = response.json["result"]
 
-    assert {"latest_version", "supported_project_version", "minimum_api_version", "maximum_api_version"} == set(
-        data.keys()
-    )
+    assert {
+        "latest_version",
+        "cli_version",
+        "supported_project_version",
+        "minimum_api_version",
+        "maximum_api_version",
+    } == set(data.keys())
     assert __version__ == data["latest_version"]
     assert SUPPORTED_PROJECT_VERSION == data["supported_project_version"]
     assert MINIMUM_VERSION.name == data["minimum_api_version"]
@@ -38,9 +42,13 @@ def test_version(svc_client):
     assert "result" in response.json
     data = response.json["result"]
 
-    assert {"latest_version", "supported_project_version", "minimum_api_version", "maximum_api_version"} == set(
-        data.keys()
-    )
+    assert {
+        "latest_version",
+        "cli_version",
+        "supported_project_version",
+        "minimum_api_version",
+        "maximum_api_version",
+    } == set(data.keys())
     assert __version__ == data["latest_version"]
     assert SUPPORTED_PROJECT_VERSION == data["supported_project_version"]
     assert MINIMUM_VERSION.name == data["minimum_api_version"]
@@ -50,9 +58,13 @@ def test_version(svc_client):
     assert "result" in response.json
     data = response.json["result"]
 
-    assert {"latest_version", "supported_project_version", "minimum_api_version", "maximum_api_version"} == set(
-        data.keys()
-    )
+    assert {
+        "latest_version",
+        "cli_version",
+        "supported_project_version",
+        "minimum_api_version",
+        "maximum_api_version",
+    } == set(data.keys())
     assert __version__ == data["latest_version"]
     assert SUPPORTED_PROJECT_VERSION == data["supported_project_version"]
     assert MINIMUM_VERSION.name == data["minimum_api_version"]
@@ -62,9 +74,13 @@ def test_version(svc_client):
     assert "result" in response.json
     data = response.json["result"]
 
-    assert {"latest_version", "supported_project_version", "minimum_api_version", "maximum_api_version"} == set(
-        data.keys()
-    )
+    assert {
+        "latest_version",
+        "cli_version",
+        "supported_project_version",
+        "minimum_api_version",
+        "maximum_api_version",
+    } == set(data.keys())
     assert __version__ == data["latest_version"]
     assert SUPPORTED_PROJECT_VERSION == data["supported_project_version"]
     assert MINIMUM_VERSION.name == data["minimum_api_version"]
@@ -74,9 +90,13 @@ def test_version(svc_client):
     assert "result" in response.json
     data = response.json["result"]
 
-    assert {"latest_version", "supported_project_version", "minimum_api_version", "maximum_api_version"} == set(
-        data.keys()
-    )
+    assert {
+        "latest_version",
+        "cli_version",
+        "supported_project_version",
+        "minimum_api_version",
+        "maximum_api_version",
+    } == set(data.keys())
     assert __version__ == data["latest_version"]
     assert SUPPORTED_PROJECT_VERSION == data["supported_project_version"]
     assert MINIMUM_VERSION.name == data["minimum_api_version"]